(for the third time) allows macros to reference enclosing `this`
In simple cases, when a scala.reflect macro is directly translated into
an inline/meta combo, we would be fine with either c.macroApplication or
meta { case macroApplication => ... }.

Unfortunately, in the case of multiple nested inline methods, it won't
be possible to access prefixes of outer inline calls. One way of addressing
this problem would be exposing c.macroApplicationStack, but I'd like to
avoid exposing non-local information as much as possible, so I started
to look for an alternative.

I think we could get away with allowing people to use `this`, which would
let them reference call prefixes. If they need trees, they could always
do `this.defn.rhs` or something like that. That's suboptimal, but at least
it brings the status of the problem from "impossible" to "unpleasant".