scala.meta.Expr => scala.meta.Term
We can't immediately rename Term to Expr in scala.meta, because that would
mean that we'd have to live with things like Expr.Ref and Expr.Param which
look suboptimal. I'd like to ponder better names for these concepts
(because as of late I came to like the name "scala.meta.Expr"), but we don't
have time for that at the moment.

Therefore, I'm undoing the speculative renaming that's already been applied to the spec
until we're ready to go ahead with the renaming in scala.meta.